Adding NEST 3.0 features, removing functionality deprecated in NEST 2.x.

+++ nestkernel/music_event_handler.cpp
@@ -71,7 +71,7 @@ MusicEventHandler::~MusicEventHandler()
 }
 
 void
-MusicEventHandler::register_channel( int channel, nest::Node* mp )
+MusicEventHandler::register_channel( size_t channel, nest::Node* mp )
 {
   if ( channel >= channelmap_.size() )
   {
